Add .gitignore, update README, and refactor game logic
- Created a .gitignore file to exclude workspace and config files.
- Updated README for clarity on project purpose.
- Refactored Makefile for consistency and readability.
- Modified Config and Brick classes to improve data handling.
- Enhanced Game class with new level loading and endgame creation features.
- Improved collision detection and game mechanics.
- Added support for saving and loading endgame configurations.
- Introduced new level files for better game structure.

+++ breakout_game/src/Config.h
@@ -5,7 +5,7 @@
 
 struct Config {
     std::string filename;
-    int ballSpeed;
+    double ballSpeed;
     int randomSeed;
     int initialLevel;
